\name{CensMFM-package}
\alias{CensMFM-package}
\alias{CensMFM}
\docType{package}
\title{
\packageTitle{CensMFM}
}
\description{
\packageDescription{CensMFM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CensMFM}
\packageIndices{CensMFM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CensMFM}

Maintainer: \packageMaintainer{CensMFM}
}
\references{
Cabral, C. R. B., Lachos, V. H., & Prates, M. O. (2012). Multivariate mixture modeling using skew-normal independent distributions. Computational Statistics & Data Analysis, 56(1), 126-142.

Prates, M. O., Lachos, V. H., & Cabral, C. (2013). mixsmsn: Fitting finite mixture of scale mixture of skew-normal distributions. Journal of Statistical Software, 54(12), 1-20.

C.E. Galarza, L.A. Matos, D.K. Dey & V.H. Lachos. (2019) On Moments of Folded and Truncated Multivariate Extended Skew-Normal Distributions. Technical report. ID 19-14. University of Connecticut.

F.H.C. de Alencar, C.E. Galarza, L.A. Matos & V.H. Lachos. (2019) Finite Mixture Modeling of Censored and Missing Data Using the Multivariate Skew-Normal Distribution. echnical report. ID 19-31. University of Connecticut.
}

\keyword{Multivariate distributions}% use one of  RShowDoc("KEYWORDS")
\keyword{Skew_normal}% __ONLY ONE__ keyword per line
\keyword{Censored data}
\keyword{Missing data}
\keyword{Finite Mixture}
\seealso{
\code{\link{fit.FMMSNC}}, \code{\link{rMSN}}, \code{\link{rMMSN}} and \code{\link{rMMSN.contour}}
}
